CNJR-2493: Allow authenticators to be configured
Previously conjur-ecs deploy did not allow a user to specify
authenticators that should be enabled. This commit adds an
environment variable (CONJUR_AUTHENTICATORS) which is read
by the prepare script, and passed to cloudformation via the
params file. The cloudformation template then sets this variable
as an environment entry for the conjur container so conjur
can read the list of allowed authenticators.